model/requirements.txt
gradio
ultralytics
torch
pillow
numpy
